backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-allauth>=0.61
dj-rest-auth>=5.0
channels>=4.0
daphne>=4.0
Pillow>=10.0
requests>=2.31
psycopg2-binary>=2.9
